requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-spectacular==0.22.0
frozenlist==1.3.0
idna==3.3
inflection==0.5.1
jsonschema==4.4.0
multidict==6.0.2
psycopg2==2.9.3
PyJWT==2.3.0
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
yarl==1.7.2
